Move autocompletion trigger to extended-codemirror
